+++ src/main/java/com/znczCxtc/service/DingDanService.java
@@ -37,9 +37,9 @@ public interface DingDanService {
 
 	DingDan getByCphJL(String cph);
 
-	int getIdByDdh(String ddh);
+	long getIdByDdh(String ddh);
 
-	boolean checkIfExistByIdCph(Integer id, String cph);
+	boolean checkIfExistByIdCph(Long id, String cph);
 
 	String createDdhByDateYMD();
